libs/mng_schedule/pyproject.toml
[project]
name = "mng-schedule"
version = "0.1.0"
description = "Schedule command plugin for mng - schedule remote invocations of mng commands"
requires-python = ">=3.11"
dependencies = [
    "mng",
    "click-option-group>=0.5.6",
    "imbue-common",
]

[project.entry-points.mng]
schedule = "imbue.mng_schedule.plugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["imbue"]

[tool.uv.sources]
mng = { workspace = true }
imbue-common = { workspace = true }

[tool.pyright]
venvPath = "../.."
venv = ".venv"
pythonVersion = "3.11"
strict = ["**/*.py"]

[tool.pytest.ini_options]
timeout_func_only = true
pythonpath = ["."]
testpaths = ["."]
addopts = [
  "--cov=imbue.mng_schedule",
  "--cov-report=term-missing",
  "--cov-report=html:.test_output/htmlcov",
  "--cov-report=xml:.test_output/coverage.xml",
  "--cov-fail-under=30",
  "--durations=20",
  "--timeout=10",
  "--timeout-method=signal",
  "-m not acceptance and not release",
  "--slow-tests-to-file",
  "--coverage-to-file",
]
